+++ 1.21.6/Common/src/main/java/fuzs/puzzleslib/api/event/v1/entity/living/LivingDropsCallback.java
@@ -3,7 +3,6 @@ package fuzs.puzzleslib.api.event.v1.entity.living;
 import fuzs.puzzleslib.api.event.v1.core.EventInvoker;
 import fuzs.puzzleslib.api.event.v1.core.EventResult;
 import net.minecraft.world.damagesource.DamageSource;
-import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.item.ItemEntity;
 
@@ -15,20 +14,17 @@ public interface LivingDropsCallback {
     /**
      * Called right before drops from a killed entity are spawned in the world.
      * <p>
-     * This allows to change the loot that is dropped.
-     * <p>
-     * The looting level can be obtained via
-     * {@link fuzs.puzzleslib.api.item.v2.EnchantingHelper#getMobLootingLevel(Entity, Entity, DamageSource)}.
+     * This allows changing the loot that is dropped.
      *
      * @param livingEntity the entity that has been killed
-     * @param damageSource damage source that killed the entity
-     * @param itemDrops    all drops, including equipment, so not just drops from the entity's loot table; this can be
-     *                     modified
+     * @param damageSource the damage source that killed the entity
+     * @param itemDrops    the modifiable drops; including both loot table items and equipment
      * @param recentlyHit  does this count as a player kill, meaning
      *                     {@link LivingEntity#getLastHurtByPlayerMemoryTime()} is not zero
-     * @return {@link EventResult#INTERRUPT} to prevent any drops from spawning, {@link EventResult#PASS} to allow drops
-     *         to be spawned in the world; will have the same result as {@link EventResult#INTERRUPT} if drops are
-     *         empty
+     * @return <ul>
+     *         <li>{@link EventResult#INTERRUPT INTERRUPT} to prevent any drops from spawning, equal to clearing the list of drops</li>
+     *         <li>{@link EventResult#PASS PASS} to allow drops to be spawned in the world</li>
+     *         </ul>
      */
     EventResult onLivingDrops(LivingEntity livingEntity, DamageSource damageSource, Collection<ItemEntity> itemDrops, boolean recentlyHit);
 }