+++ 1.21.6/CHANGELOG.md
@@ -4,14 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v21.6.0-1.21.6] - 2025-06-03
+## [v21.6.0-1.21.6] - 2025-06-08
 - Update to Minecraft 1.21.6
+### Added
+- Add `CommonHelper::getMinecraftServer`
+- Add `CodecExtras::fromEnum`
+- Add `RegistryManager::registerEnchantmentEffectComponentType`
+- Add `ResourceKeyHelper::getResourceLocation`
+- Add many enchantment value effect bonus helper methods to `EnchantingHelper`
 ### Changed
 - Migrate `GuiLayersContext` to use custom layer implementation
 - Rework `FogEvents` to account for vanilla changes
 - Simplify `RenderLevelEvents` implementation as `RenderLevelCallback`
 - Greatly simplify `ContainerSerializationHelper`
 - Replace `AnvilEvents` with a single `CreateAnvilResultCallback`
+- Move some methods from `LookupHelper` to dedicated classes
 ### Removed
 - Remove `fuzs.puzzleslib.api.capability.v3`
 - Remove `fuzs.puzzleslib.api.network.v3`
@@ -21,3 +28,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove `ExtendedMenuSupplier`
 - Remove `EnchantingHelper::isBookEnchantable`
 - Remove `BlockEvents$FarmlandTrample`
+- Remove defaulted mutable value classes from public api
+- Remove `ComputeEnchantedLootBonusCallback`
+- Remove `EnchantingHelper::getMobLootingLevel`